FlaskProject/requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.6
python-dotenv==1.0.0